Fix pendingTask count not tracking task rejection
Fixes issue with TimedScheduler where tasks rejected by underlying
scheduler would not be removed from pendingTaks count. Idea is to
catch any RejectedExecutionException thrown when scheduling the
task, stop its pending sample to remove it from the pending count, and
rethrow the error so caller knows that the task was rejected.

Fixes #3642