[doc] Improve Mono#share javadoc for cancelling
Similarly to Flux#share, Mono#share also cancels the source when all
Subscribers have cancelled. This change improves the documentation.

Following #2680 and #2756 there exists a misalignment in the javadoc for
Mono#share method. Since cancelling the source is a fact, the
javadoc is now improved instead of changing the behaviour.

Resolves #3740