[build] Upgrade 3.4.x to Gradle 8.6 (#3745)
This is a backport of Gradle 8 upgrade that was introduced in 3.6.x. It
aims to ease development and bring consistency.

It also involves upgrade of the Kotlin compiler to 1.8, however still
targetting 1.3. In 1.9 that support has been removed, so 1.8 will remain
in 3.4. In 3.5 there are no kotlin extensions, they are part of
reactior-kotlin-extensions project.

---------

Co-authored-by: Oleh Dokuka <5380167+OlegDokuka@users.noreply.github.com>