[build] Parallelize tests (#3764)
Introduces parallel gradle execution of all the modules. Also, runs
JUnit tests in parallel.

With this change the JUnit tests run faster. This removes the need for
separating "slow" tests in CI and during publishing. Also, because the
modules are executed in parallel, the overall time is better. As an
expected outcome, our Gradle workflow files should be simplified.

Resolves #2270