Rework BufferTimeout with fair backpressure (#3634)
The fair backpressure variant of the bufferTimeout operator has been
reworked to use a state machine with a minimum number of volatile
variables eliminating potential data races, such as skipping the
delivery when onNext and timeout happen concurrently or cancellation
happens while onNext is delivered, etc.

The changes include an improved JCstress suite to cover discarded values
and more racing scenarios. A serial JMH benchmark can also be exercised.
Also, the implementation contains comments to aid in future
understanding of the algorithm and a set of utility methods which can
form a basis for a generic state machine utility toolset.