Call init on delegate from TimedScheduler
The TimedScheduler class doesn't currently provide an override for the
`init()` method, which means it calls the default implementation on the
interface that delegates to `start()`. This works fine for most
schedulers, since they have a valid implementation of `start()`. However,
the newer BoundedElasticThreadPerTaskScheduler throws an error for
`start()`, so wrapping it in a TimedScheduler causes it to crash
immediately when `init()` gets called.

We should call the wrapped scheduler's `init()` method instead, which
allows users to get metrics for their virtual thread schedulers.