Call init() on delegate from TimedScheduler (#3876)
The `TimedScheduler` class doesn't currently provide an override for the
`init()` method, which means it calls the default implementation on the
interface that delegates to `start()`. Most existing schedulers have an
implementation of `start()` and it's not a problem for them. However,
the newer `BoundedElasticThreadPerTaskScheduler` throws an error for
`start()`, so wrapping it in a `TimedScheduler` causes it to crash
immediately when `init()` gets called.

With this change, the wrapped scheduler's `init()` method is called
instead and allows users to get metrics for the virtual thread
schedulers.

---------

Co-authored-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>