Fix `SwitchOnFirst` hanging subscription
In a situation where `SwitchOnFirstMain` is subscribed to after the
operator itself has terminated, the expectation is that either an error
is propagated due to cancellation of the source or the first actual
subscription succeeds. Due to a regression introduced in #2794 the
subscription would hang with no termination. This change addresses the
issue and the subscrption properly terminates.

Related to #3936.

Signed-off-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>