Interrupt VirtualThread task only from other Thread
When `BoundedElasticScheduler` processes a task, the task can be disposed.
In the regular case the Thread executing the task will only be
interrupted when the disposing Thread is different.

With the Virtual Thread-based implementation this behaviour should be
similar. Since it was not, this change aligns the new implementation to
be similar in that regard.

Fixes #3948

Signed-off-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>