[build] Compile with recent javac but target JDK 8
Using Gradle toolchains and mrjar plugin it is difficult to use the
latest JDK to compile against JDK 8 but with the latest compiler. This
change enhances the JavaConventions internal plugin to allow taking
advantage of more recent features but compiling appropriate byte code
levels. This also means the java21 and related version-specific tasks
are compile with the relevant version and tests are run with the
required version.

toolchains.gradle configuration was removed as it was not used and made
things more difficult at this point by colliding with the above
enhancement when used.