[test] Relax Flux#switchMap test2596 to avoid timeouts (#4004)
The test times out frequently on CI. If there is a genuine issue, the
test will still time out due to an infinite loop condition that this
test tries to detect. Reducing the number of executions and shutting
down schedulers should reduce the time taken in valid scenarios and the
flakiness should be minimized.

Signed-off-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>