Adjust reactor.util.Logger vararg nullability (#4098)
With JSpecify annotations, array and varargs nullability checks are more
expressive and provide more out-of-the-box strictness. The
reactor.util.Logger interface has relaxed requirements regarding varargs
parameters when logging so that is now reflected using the type-use
annotations syntax.  Additionally, the TestLogger was adjusted to avoid
NPE when null varargs elements are provided.

Resolves #4095

Signed-off-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>